chore: Remove frontend/ from monorepo.
Removed the frontend codebase from the monorepo in preparation for
moving it to a separate repository, in order to improve separation of
concerns and independent versioning and deployment.

The repository will now be dedicated to the backend API, and all code
related to it, only. No backend functionality was modified.